Module abi_stable::abi_stability

types and traits related to abi stability.

Re-exports

pub use self::get_static_equivalent::GetStaticEquivalent;

pub use self::get_static_equivalent::GetStaticEquivalent_;

pub use self::stable_abi_trait::AbiConsts;

pub use self::stable_abi_trait::PrefixStableAbi;

pub use self::stable_abi_trait::StableAbi;

pub use self::extra_checks::ExtraChecks;

pub use self::extra_checks::TypeChecker;

Modules

extra_checks

Contains items for adding checks to individual types.

get_static_equivalent

Contains the GetStaticEquivalent_ trait and related items.

stable_abi_trait

Where the StableAbi trait is declared,as well as related types/traits.

Structs

ConstGeneric

A trait object used to check equality between const generic parameters.

Functions

check_layout_compatibility

Checks that the layout of interface is compatible with implementation,