Module abi_stable::abi_stability::extra_checks::ExtraChecks_trait

This module is generated by the #[sabi_trait] attribute on ExtraChecks

Structs

ExtraChecks_Interface

A marker type describing the traits that are required when constructing ExtraChecks_TO,and are then implemented by it, by implementing the InterfaceType trait.

ExtraChecks_TO

The trait object for ExtraChecks.

Traits

ExtraChecks

Allows defining extra checks for a type.

Type Aliases

ExtraChecks_Backend

An alias for the underlying implementation of ExtraChecks_TO.

ExtraChecks_CTO

A type alias for the const-constructible ExtraChecks_TO.