abi_stable::abi_stability

Module stable_abi_trait

Where the StableAbi trait is declared,as well as related types/traits.

Structs

AbiConsts

Contains constants equivalent to the associated types in StableAbi.

SabiUnsafeOpaqueField

Allows one to ensure that a T implements StableAbi, while storing an opaque layout instead of <T as StableAbi>::LAYOUT.

UnsafeOpaqueField

Allows one to create the TypeLayout for any type T, by pretending that it is a primitive type.

Traits

PrefixStableAbi

A type that only has a stable layout when a PrefixRef to it is used.

StableAbi

Represents a type whose layout is stable.

Functions

get_prefix_field_type_layout

Retrieves the TypeLayout of T: PrefixStableAbi,

get_type_layout

Retrieves the TypeLayout of T: StableAbi,