Struct abi_stable::erased_types::DynTrait

#[repr(C)]
pub struct DynTrait<'borr, P, I, EV = ()>where
    P: GetPointerKind,{ /* private fields */ }

DynTrait implements ffi-safe trait objects, for a selection of traits.

Passing opaque values around with DynTrait<_>

One can pass non-StableAbi types around by using type erasure, using this type.

It generally looks like DynTrait<'borrow, Pointer<()>, Interface>, where:

'borrow is the borrow that the type that was erased had.

Pointer is a pointer type that implements AsPtr.

Interface is an InterfaceType, which describes what traits are required when constructing the DynTrait<_> and which ones it implements.

Construction

To construct a DynTrait<_> one can use these associated functions:

from_value: Can be constructed from the value directly.Requires a 'static value.

from_ptr Can be constructed from a pointer of a value.Requires a 'static value.

from_borrowing_value: Can be constructed from the value directly.Cannot downcast the DynTrait afterwards.

from_borrowing_ptr Can be constructed from a pointer of a value.Cannot downcast the DynTrait afterwards.

DynTrait uses the impls of the value in methods, which means that the pointer itself does not have to implement those traits,

Trait object

DynTrait<'borrow, Pointer<()>, Interface> can be used as a trait object for any combination of the traits listed below.

These are the traits:

Send

Sync

Unpin

Iterator

DoubleEndedIterator

std::fmt::Write

std::io::Write

std::io::Seek

std::io::Read

std::io::BufRead

Clone

Display

Debug

std::error::Error

Default: Can only be called as an inherent method.

Eq

PartialEq

Ord

PartialOrd

Hash

serde::Deserialize: first deserializes from a string, and then calls the objects’ Deserialize impl.

serde::Serialize: first calls the objects’ Deserialize impl, then serializes that as a string.

Deconstruction

DynTrait<_> can then be unwrapped into a concrete type, within the same dynamic library/executable that constructed it, using these (fallible) conversion methods:

downcast_into: Unwraps into a pointer to T.Requires T:'static.

downcast_as: Unwraps into a &T.Requires T:'static.

downcast_as_mut: Unwraps into a &mut T.Requires T:'static.

DynTrait cannot be converted back if it was created using DynTrait::from_borrowing_*.

Passing DynTrait between dynamic libraries

Passing DynTrait between dynamic libraries (as in between the dynamic libraries directly loaded by the same binary/dynamic library) may cause the program to panic at runtime with an error message stating that the trait is not implemented for the specific interface.

This can only happen if you are passing DynTrait between dynamic libraries, or if DynTrait was instantiated in the parent passed to a child, a DynTrait instantiated in a child dynamic library passed to the parent should not cause a panic, it would be a bug.

        binary
  _________|___________
lib0      lib1      lib2
  |         |         |
lib00    lib10      lib20

In this diagram passing a DynTrait constructed in lib00 to anything other than the binary or lib0 will cause the panic to happen if:

The InterfaceType requires extra traits in the version of the Interface that lib1 and lib2 know about (that the binary does not require).

lib1 or lib2 attempt to call methods that require the traits that were added to the InterfaceType, in versions of that interface that only they know about.

Examples

In the Readme

The primary example using DynTrait<_> is in the readme.

Readme is in the repository for this crate, crates.io, lib.rs.

Serialization/Deserialization

The DeserializeDyn and SerializeType demonstrate how DynTrait can be de/serialized.

Comparing DynTraits

This is only possible if the erased types don’t contain borrows, and they are not constructed using DynTrait::from_borrowing_* methods.

DynTraits wrapping different pointer types can be compared with each other, it simply uses the values’ implementation of PartialEq.

use abi_stable::{
    erased_types::interfaces::PartialEqInterface,
    std_types::{RArc, RBox},
    DynTrait, RMut, RRef,
};

{
    // `DynTrait`s constructed from `&` are `DynTrait<'_, RRef<'_, ()>, _>`
    // since `&T` can't soundly be transmuted back and forth into `&()`
    let left: DynTrait<'static, RRef<'_, ()>, PartialEqInterface> =
        DynTrait::from_ptr(&100);

    let mut n100 = 100;
    // `DynTrait`s constructed from `&mut` are `DynTrait<'_, RMut<'_, ()>, _>`
    // since `&mut T` can't soundly be transmuted back and forth into `&mut ()`
    let right: DynTrait<'static, RMut<'_, ()>, PartialEqInterface> =
        DynTrait::from_ptr(&mut n100).interface(PartialEqInterface);

    assert_eq!(left, right);
}
{
    let left: DynTrait<'static, RBox<()>, PartialEqInterface> =
        DynTrait::from_value(200);

    let right: DynTrait<'static, RArc<()>, _> =
        DynTrait::from_ptr(RArc::new(200)).interface(PartialEqInterface);

    assert_eq!(left, right);
}

Writing to a DynTrait

This is an example of using the write!() macro with DynTrait.

use abi_stable::{erased_types::interfaces::FmtWriteInterface, DynTrait, RMut};

use std::fmt::Write;

let mut buffer = String::new();

let mut wrapped: DynTrait<'static, RMut<'_, ()>, FmtWriteInterface> =
    DynTrait::from_ptr(&mut buffer);

write!(wrapped, "Foo").unwrap();
write!(wrapped, "Bar").unwrap();
write!(wrapped, "Baz").unwrap();

drop(wrapped);

assert_eq!(&buffer[..], "FooBarBaz");

Iteration

Using DynTrait as an Iterator and DoubleEndedIterator.

use abi_stable::{erased_types::interfaces::DEIteratorInterface, DynTrait};

let mut wrapped = DynTrait::from_value(0..=10).interface(DEIteratorInterface::NEW);

assert_eq!(
    wrapped.by_ref().take(5).collect::<Vec<_>>(),
    vec![0, 1, 2, 3, 4]
);

assert_eq!(wrapped.rev().collect::<Vec<_>>(), vec![10, 9, 8, 7, 6, 5]);

Making pointers compatible with DynTrait

To make pointers compatible with DynTrait, they must imlement the abi_stable::pointer_trait::{ GetPointerKind, AsPtr, AsMutPtr, CanTransmuteElement} traits as shown in the example.

GetPointerKind should generally be implemented with type Kind = PK_SmartPointer. The exception is in the case that it is a #[repr(transparent)] wrapper around a RRef/RMut/*const T/*mut T/NonNull, in which case it should implement GetPointerKind<Kind = PK_Reference> (when it has shared reference semantics) or GetPointerKind<Kind = PK_MutReference> (when it has mutable reference semantics).

Example

This is an example of a newtype wrapping an RBox, demonstrating that the pointer type doesn’t have to implement the traits in the InterfaceType, it’s the value it points to.

     
use abi_stable::DynTrait;

fn main() {
    let lines = "line0\nline1\nline2";
    let mut iter = NewtypeBox::new(lines.lines());

    // The type annotation here is just to show the type, it's not necessary.
    let mut wrapper: DynTrait<'_, NewtypeBox<()>, IteratorInterface> =
        DynTrait::from_borrowing_ptr(iter);

    // You can clone the DynTrait!
    let clone = wrapper.clone();

    assert_eq!(wrapper.next(), Some("line0"));
    assert_eq!(wrapper.next(), Some("line1"));
    assert_eq!(wrapper.next(), Some("line2"));
    assert_eq!(wrapper.next(), None);

    assert_eq!(
        clone.rev().collect::<Vec<_>>(),
        vec!["line2", "line1", "line0"],
    )
}

#[repr(C)]
#[derive(StableAbi)]
#[sabi(impl_InterfaceType(
    Sync,
    Send,
    Iterator,
    DoubleEndedIterator,
    Clone,
    Debug
))]
pub struct IteratorInterface;

impl<'a> IteratorItem<'a> for IteratorInterface {
    type Item = &'a str;
}

/////////////////////////////////////////

use std::ops::{Deref, DerefMut};

use abi_stable::{
    erased_types::IteratorItem,
    pointer_trait::{
        AsMutPtr, AsPtr, CanTransmuteElement, GetPointerKind, PK_SmartPointer,
    },
    std_types::RBox,
    type_level::bools::True,
    InterfaceType, StableAbi,
};

#[repr(transparent)]
#[derive(Default, Clone, StableAbi)]
pub struct NewtypeBox<T> {
    box_: RBox<T>,
}

impl<T> NewtypeBox<T> {
    pub fn new(value: T) -> Self {
        Self {
            box_: RBox::new(value),
        }
    }
}

unsafe impl<T> GetPointerKind for NewtypeBox<T> {
    // This is a smart pointer because `RBox` is one.
    type Kind = PK_SmartPointer;
    type PtrTarget = T;
}

// safety: Does not create an intermediate `&T` to get a pointer to `T`.
unsafe impl<T> AsPtr for NewtypeBox<T> {
    fn as_ptr(&self) -> *const T {
        self.box_.as_ptr()
    }
}

// safety: Does not create an intermediate `&mut T` to get a pointer to `T`
unsafe impl<T> AsMutPtr for NewtypeBox<T> {
    fn as_mut_ptr(&mut self) -> *mut T {
        self.box_.as_mut_ptr()
    }
}

// safety:
// NewtypeBox is safe to transmute, because RBox (the pointer type it wraps)
// is safe to transmute
unsafe impl<T, O> CanTransmuteElement<O> for NewtypeBox<T> {
    type TransmutedPtr = NewtypeBox<O>;

    unsafe fn transmute_element_(self) -> Self::TransmutedPtr {
        let box_: RBox<O> = self.box_.transmute_element_();
        NewtypeBox { box_ }
    }
}

Implementations

impl<I> DynTrait<'static, RBox<()>, I>

pub fn from_value<T>(object: T) -> Selfwhere
T: 'static,
VTable_Ref<'static, RBox<()>, I>: MakeVTable<'static, T, RBox<T>, TD_CanDowncast>,

Constructs the DynTrait<_> from a type that doesn’t borrow anything.

Example

use abi_stable::{
    erased_types::interfaces::DebugDisplayInterface, std_types::RBox, DynTrait,
};

// DebugDisplayInterface is `Debug + Display + Sync + Send`
let to: DynTrait<'static, RBox<()>, DebugDisplayInterface> =
    DynTrait::from_value(3u8);

assert_eq!(format!("{}", to), "3");
assert_eq!(format!("{:?}", to), "3");

impl<P, I> DynTrait<'static, P, I>where
P: GetPointerKind,

pub fn from_ptr<OrigPtr>(object: OrigPtr) -> Selfwhere
OrigPtr: GetPointerKind + CanTransmuteElement<(), TransmutedPtr = P>,
OrigPtr::PtrTarget: 'static,
VTable_Ref<'static, P, I>: MakeVTable<'static, OrigPtr::PtrTarget, OrigPtr, TD_CanDowncast>,

Constructs the DynTrait<_> from a pointer to a type that doesn’t borrow anything.

Example

use abi_stable::{
    erased_types::interfaces::DebugDisplayInterface,
    std_types::{RArc, RBox},
    DynTrait, RMut, RRef,
};

// Constructing a DynTrait from a `&T`
{
    // `DynTrait`s constructed from `&` are `DynTrait<'_, RRef<'_, ()>, _>`
    // since `&T` can't soundly be transmuted back and forth into `&()`
    let rref: DynTrait<'static, RRef<'_, ()>, DebugDisplayInterface> =
        DynTrait::from_ptr(&21i32);

    assert_eq!(format!("{:?}", rref), "21");
    assert_eq!(format!("{}", rref), "21");
}
// Constructing a DynTrait from a `&mut T`
{
    let mmut = &mut "hello";
    // `DynTrait`s constructed from `&mut` are `DynTrait<'_, RMut<'_, ()>, _>`
    // since `&mut T` can't soundly be transmuted back and forth into `&mut ()`
    let rmut: DynTrait<'static, RMut<'_, ()>, DebugDisplayInterface> =
        DynTrait::from_ptr(mmut).interface(DebugDisplayInterface);

    assert_eq!(format!("{:?}", rmut), r#""hello""#);
    assert_eq!(format!("{}", rmut), "hello");
}
// Constructing a DynTrait from a `RBox<T>`
{
    let boxed: DynTrait<'static, RBox<()>, DebugDisplayInterface> =
        DynTrait::from_ptr(RBox::new(false));

    assert_eq!(format!("{:?}", boxed), "false");
    assert_eq!(format!("{}", boxed), "false");
}
// Constructing a DynTrait from an `RArc<T>`
{
    let arc: DynTrait<'static, RArc<()>, DebugDisplayInterface> =
        DynTrait::from_ptr(RArc::new(30u32)).interface(DebugDisplayInterface);

    assert_eq!(format!("{:?}", arc), "30");
}

impl<'borr, I> DynTrait<'borr, RBox<()>, I>

pub fn from_borrowing_value<T>(object: T) -> Selfwhere
T: 'borr,
VTable_Ref<'borr, RBox<()>, I>: MakeVTable<'borr, T, RBox<T>, TD_Opaque>,

Constructs the DynTrait<_> from a value with a 'borr borrow.

Cannot downcast the DynTrait afterwards.

Example

use abi_stable::{
    erased_types::interfaces::DebugDisplayInterface, std_types::RBox, DynTrait,
};

// DebugDisplayInterface is `Debug + Display + Sync + Send`
let to: DynTrait<'static, RBox<()>, DebugDisplayInterface> =
    DynTrait::from_borrowing_value(3u8);

assert_eq!(format!("{}", to), "3");
assert_eq!(format!("{:?}", to), "3");

// `DynTrait`s constructed using the `from_borrowing_*` constructors
// can't be downcasted.
assert_eq!(to.downcast_as::<u8>().ok(), None);

impl<'borr, P, I> DynTrait<'borr, P, I>where
P: GetPointerKind,

pub fn from_borrowing_ptr<OrigPtr>(object: OrigPtr) -> Selfwhere
OrigPtr: GetPointerKind + 'borr + CanTransmuteElement<(), TransmutedPtr = P>,
OrigPtr::PtrTarget: 'borr,
VTable_Ref<'borr, P, I>: MakeVTable<'borr, OrigPtr::PtrTarget, OrigPtr, TD_Opaque>,

Constructs the DynTrait<_> from a pointer to the erased type with a 'borr borrow.

Cannot downcast the DynTrait afterwards.

Example

use abi_stable::{
    erased_types::interfaces::DebugDisplayInterface,
    std_types::{RArc, RBox},
    DynTrait, RMut, RRef,
};

// Constructing a DynTrait from a `&T`
{
    // `DynTrait`s constructed from `&` are `DynTrait<'_, RRef<'_, ()>, _>`
    // since `&T` can't soundly be transmuted back and forth into `&()`
    let rref: DynTrait<'_, RRef<'_, ()>, DebugDisplayInterface> =
        DynTrait::from_borrowing_ptr(&34i32);

    assert_eq!(format!("{:?}", rref), "34");
    assert_eq!(format!("{}", rref), "34");
}
// Constructing a DynTrait from a `&mut T`
{
    let mmut = &mut "world";
    // `DynTrait`s constructed from `&mut` are `DynTrait<'_, RMut<'_, ()>, _>`
    // since `&mut T` can't soundly be transmuted back and forth into `&mut ()`
    let rmut: DynTrait<'_, RMut<'_, ()>, DebugDisplayInterface> =
        DynTrait::from_borrowing_ptr(mmut).interface(DebugDisplayInterface);

    assert_eq!(format!("{:?}", rmut), r#""world""#);
    assert_eq!(format!("{}", rmut), "world");
}
// Constructing a DynTrait from a `RBox<T>`
{
    let boxed: DynTrait<'_, RBox<()>, DebugDisplayInterface> =
        DynTrait::from_borrowing_ptr(RBox::new(true));

    assert_eq!(format!("{:?}", boxed), "true");
    assert_eq!(format!("{}", boxed), "true");
}
// Constructing a DynTrait from an `RArc<T>`
{
    let arc: DynTrait<'_, RArc<()>, _> =
        DynTrait::from_borrowing_ptr(RArc::new('a')).interface(DebugDisplayInterface);

    assert_eq!(format!("{:?}", arc), "'a'");
    assert_eq!(format!("{}", arc), "a");
}

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
P: AsPtr<PtrTarget = ()>,

pub fn with_extra_value<OrigPtr, Downcasting>(
ptr: OrigPtr,
extra_value: EV
) -> DynTrait<'borr, P, I, EV> where
OrigPtr: GetPointerKind + CanTransmuteElement<(), TransmutedPtr = P>,
OrigPtr::PtrTarget: 'borr,
VTable_Ref<'borr, P, I>: MakeVTable<'borr, OrigPtr::PtrTarget, OrigPtr, Downcasting>,

Constructs an DynTrait from an erasable pointer and an extra value.

Example

use abi_stable::{
    erased_types::{interfaces::DebugDisplayInterface, TD_Opaque},
    DynTrait, RRef,
};

// DebugDisplayInterface is `Debug + Display + Sync + Send`
let to: DynTrait<'static, RRef<()>, DebugDisplayInterface, usize> =
    DynTrait::with_extra_value::<_, TD_Opaque>(&55u8, 100usize);

assert_eq!(format!("{}", to), "55");
assert_eq!(format!("{:?}", to), "55");

assert_eq!(to.sabi_extra_value(), &100);

impl<'borr, 'a, I, EV> DynTrait<'borr, RRef<'a, ()>, I, EV>

pub const fn from_const<T, Downcasting>(
ptr: &'a T,
can_it_downcast: Downcasting,
extra_value: EV
) -> Selfwhere
VTable_Ref<'borr, RRef<'a, ()>, I>: MakeVTable<'borr, T, &'a T, Downcasting>,
T: 'borr,

This function allows constructing a DynTrait in a constant/static.

Parameters

ptr: a reference to the value.

can_it_downcast can be either:

TD_CanDowncast: Which allows the trait object to be downcasted, requires that the value implements any.

TD_Opaque: Which does not allow the trait object to be downcasted.

extra_value: This is used by #[sabi_trait] trait objects to store their vtable inside DynTrait.

Example

use abi_stable::{
    erased_types::{
        interfaces::DebugDisplayInterface, DynTrait, TD_Opaque,
    },
    sabi_types::RRef,
};

static STRING: &str = "What the heck";

static DYN: DynTrait<'static, RRef<'static, ()>, DebugDisplayInterface, ()> =
    DynTrait::from_const(&STRING, TD_Opaque, ());

fn main() {
    assert_eq!(format!("{}", DYN), format!("{}", STRING));
    assert_eq!(format!("{:?}", DYN), format!("{:?}", STRING));
}

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
P: GetPointerKind,

pub const fn interface(self, _interface: I) -> Selfwhere
I: InterfaceType,

For inferring the I type parameter.

impl<P, I, EV> DynTrait<'static, P, I, EV>where
P: GetPointerKind,

pub fn sabi_is_same_type<Other, I2, EV2>(
&self,
other: &DynTrait<'static, Other, I2, EV2>
) -> boolwhere
I2: InterfaceType,
Other: GetPointerKind,

Allows checking whether 2 DynTrait<_>s have a value of the same type.

Notes:

Types from different dynamic libraries/executables are never considered equal.

DynTraits constructed using DynTrait::from_borrowing_* are never considered to wrap the same type.

impl<'borr, P, I, EV> DynTrait<'borr, P, I, PrefixRef<EV>>where
P: GetPointerKind,

pub const fn sabi_et_vtable(&self) -> PrefixRef<EV>

A vtable used by #[sabi_trait] derived trait objects.

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
P: GetPointerKind,

pub const fn sabi_extra_value(&self) -> &EV

Gets access to the extra value that was stored in this DynTrait in the with_extra_value constructor.

Example

use abi_stable::{erased_types::TD_Opaque, DynTrait, RRef};

let to: DynTrait<'static, RRef<()>, (), char> =
    DynTrait::with_extra_value::<_, TD_Opaque>(&55u8, 'Z');

assert_eq!(to.sabi_extra_value(), &'Z');

pub fn sabi_object_address(&self) -> usizewhere
P: AsPtr,

Returns the address of the wrapped object.

Example

use abi_stable::{erased_types::TD_Opaque, DynTrait, RRef};

let reff = &55u8;

let to: DynTrait<'static, RRef<()>, ()> = DynTrait::from_ptr(reff);

assert_eq!(to.sabi_object_address(), reff as *const _ as usize);

pub fn sabi_erased_ref(&self) -> RRef<'_, ErasedObject>where
P: AsPtr,

Gets a reference pointing to the erased object.

Example

use abi_stable::{std_types::RBox, DynTrait};

let to: DynTrait<'static, RBox<()>, ()> = DynTrait::from_value(66u8);

unsafe {
    assert_eq!(to.sabi_erased_ref().transmute_into_ref::<u8>(), &66);
}

pub fn sabi_erased_mut(&mut self) -> RMut<'_, ErasedObject>where
P: AsMutPtr,

Gets a mutable reference pointing to the erased object.

Example

use abi_stable::{std_types::RBox, DynTrait};

let mut to: DynTrait<'static, RBox<()>, ()> = DynTrait::from_value("hello");

unsafe {
    assert_eq!(
        to.sabi_erased_mut().transmute_into_mut::<&str>(),
        &mut "hello"
    );
}

pub fn sabi_as_rref(&self) -> RRef<'_, ()>where
P: AsPtr,

Gets an RRef pointing to the erased object.

Example

use abi_stable::{std_types::RBox, DynTrait};

let to: DynTrait<'static, RBox<()>, ()> = DynTrait::from_value(66u8);

unsafe {
    assert_eq!(to.sabi_as_rref().transmute_into_ref::<u8>(), &66);
}

pub fn sabi_as_rmut(&mut self) -> RMut<'_, ()>where
P: AsMutPtr,

Gets an RMut pointing to the erased object.

Example

use abi_stable::{std_types::RBox, DynTrait};

let mut to: DynTrait<'static, RBox<()>, ()> = DynTrait::from_value("hello");

unsafe {
    assert_eq!(to.sabi_as_rmut().transmute_into_mut::<&str>(), &mut "hello");
}

pub fn sabi_with_value<F, R>(self, f: F) -> Rwhere
P: OwnedPointer<PtrTarget = ()>,
F: FnOnce(MovePtr<'_, ()>) -> R,

Calls the f callback with an MovePtr pointing to the erased object.

Example

use abi_stable::{
    sabi_types::MovePtr,
    std_types::{RBox, RString, RVec},
    DynTrait,
};

let to: DynTrait<'static, RBox<()>, ()> =
    DynTrait::from_value(RVec::<u8>::from_slice(b"foobarbaz"));

let string = to.sabi_with_value(|x| unsafe {
    MovePtr::into_inner(MovePtr::transmute::<String>(x))
});

assert_eq!(string, "foobarbaz");

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
P: GetPointerKind,

pub fn downcast_into<T>(self) -> Result<P::TransmutedPtr, UneraseError<Self>>where
T: 'static,
P: CanTransmuteElement<T>,

Unwraps the DynTrait<_> into a pointer of the concrete type that it was constructed with.

T is required to not borrow anything.

Errors

This will return an error in any of these conditions:

It is called in a dynamic library/binary outside the one from which this DynTrait<_> was constructed.

The DynTrait was constructed using a from_borrowing_* method

T is not the concrete type this DynTrait<_> was constructed with.

Example

use abi_stable::{
    std_types::{RArc, RBox},
    DynTrait,
};

{
    fn to() -> DynTrait<'static, RBox<()>, ()> {
        DynTrait::from_value(b'A')
    }

    assert_eq!(to().downcast_into::<u8>().ok(), Some(RBox::new(b'A')));
    assert_eq!(to().downcast_into::<u16>().ok(), None);
}
{
    fn to() -> DynTrait<'static, RArc<()>, ()> {
        DynTrait::from_ptr(RArc::new(b'B'))
    }

    assert_eq!(to().downcast_into::<u8>().ok(), Some(RArc::new(b'B')));
    assert_eq!(to().downcast_into::<u16>().ok(), None);
}

pub fn downcast_as<T>(&self) -> Result<&T, UneraseError<&Self>>where
T: 'static,
P: AsPtr,

Unwraps the DynTrait<_> into a reference of the concrete type that it was constructed with.

T is required to not borrow anything.

Errors

This will return an error in any of these conditions:

It is called in a dynamic library/binary outside the one from which this DynTrait<_> was constructed.

The DynTrait was constructed using a from_borrowing_* method

T is not the concrete type this DynTrait<_> was constructed with.

Example

use abi_stable::{std_types::RArc, DynTrait, RMut, RRef};

{
    let to: DynTrait<'static, RRef<'_, ()>, ()> = DynTrait::from_ptr(&9u8);

    assert_eq!(to.downcast_as::<u8>().ok(), Some(&9u8));
    assert_eq!(to.downcast_as::<u16>().ok(), None);
}
{
    let mut val = 7u8;

    let to: DynTrait<'static, RMut<'_, ()>, ()> =
        DynTrait::from_ptr(&mut val);

    assert_eq!(to.downcast_as::<u8>().ok(), Some(&7));
    assert_eq!(to.downcast_as::<u16>().ok(), None);
}
{
    let to: DynTrait<'static, RArc<()>, ()> =
        DynTrait::from_ptr(RArc::new(1u8));

    assert_eq!(to.downcast_as::<u8>().ok(), Some(&1u8));
    assert_eq!(to.downcast_as::<u16>().ok(), None);
}

pub fn downcast_as_mut<T>(&mut self) -> Result<&mut T, UneraseError<&mut Self>>where
T: 'static,
P: AsMutPtr,

Unwraps the DynTrait<_> into a mutable reference of the concrete type that it was constructed with.

T is required to not borrow anything.

Errors

This will return an error in any of these conditions:

It is called in a dynamic library/binary outside the one from which this DynTrait<_> was constructed.

The DynTrait was constructed using a from_borrowing_* method

T is not the concrete type this DynTrait<_> was constructed with.

Example

use abi_stable::{std_types::RBox, DynTrait, RMut};

{
    let mut val = 7u8;

    let mut to: DynTrait<'static, RMut<'_, ()>, ()> =
        DynTrait::from_ptr(&mut val);

    assert_eq!(to.downcast_as_mut::<u8>().ok(), Some(&mut 7));
    assert_eq!(to.downcast_as_mut::<u16>().ok(), None);
}
{
    let mut to: DynTrait<'static, RBox<()>, ()> =
        DynTrait::from_ptr(RBox::new(1u8));

    assert_eq!(to.downcast_as_mut::<u8>().ok(), Some(&mut 1u8));
    assert_eq!(to.downcast_as_mut::<u16>().ok(), None);
}

pub unsafe fn unchecked_downcast_into<T>(self) -> P::TransmutedPtrwhere
P: AsPtr + CanTransmuteElement<T>,

Unwraps the DynTrait<_> into a pointer to T, without checking whether T is the type that the DynTrait was constructed with.

Safety

You must check that T is the type that DynTrait was constructed with through other means.

Example

use abi_stable::{
    std_types::{RArc, RBox},
    DynTrait,
};

unsafe {
    fn to() -> DynTrait<'static, RBox<()>, ()> {
        DynTrait::from_value(b'A')
    }

    assert_eq!(to().unchecked_downcast_into::<u8>(), RBox::new(b'A'));
}
unsafe {
    fn to() -> DynTrait<'static, RArc<()>, ()> {
        DynTrait::from_ptr(RArc::new(b'B'))
    }

    assert_eq!(to().unchecked_downcast_into::<u8>(), RArc::new(b'B'));
}

pub unsafe fn unchecked_downcast_as<T>(&self) -> &Twhere
P: AsPtr,

Unwraps the DynTrait<_> into a reference to T, without checking whether T is the type that the DynTrait was constructed with.

Safety

You must check that T is the type that DynTrait was constructed with through other means.

Example

use abi_stable::{std_types::RArc, DynTrait, RMut, RRef};

unsafe {
    let to: DynTrait<'static, RRef<'_, ()>, ()> = DynTrait::from_ptr(&9u8);

    assert_eq!(to.unchecked_downcast_as::<u8>(), &9u8);
}
unsafe {
    let mut val = 7u8;

    let to: DynTrait<'static, RMut<'_, ()>, ()> =
        DynTrait::from_ptr(&mut val);

    assert_eq!(to.unchecked_downcast_as::<u8>(), &7);
}
unsafe {
    let to: DynTrait<'static, RArc<()>, ()> =
        DynTrait::from_ptr(RArc::new(1u8));

    assert_eq!(to.unchecked_downcast_as::<u8>(), &1u8);
}

pub unsafe fn unchecked_downcast_as_mut<T>(&mut self) -> &mut Twhere
P: AsMutPtr,

Unwraps the DynTrait<_> into a mutable reference to T, without checking whether T is the type that the DynTrait was constructed with.

Safety

You must check that T is the type that DynTrait was constructed with through other means.

Example

use abi_stable::{std_types::RBox, DynTrait, RMut};

unsafe {
    let mut val = 7u8;

    let mut to: DynTrait<'static, RMut<'_, ()>, ()> =
        DynTrait::from_ptr(&mut val);

    assert_eq!(to.unchecked_downcast_as_mut::<u8>(), &mut 7);
}
unsafe {
    let mut to: DynTrait<'static, RBox<()>, ()> =
        DynTrait::from_ptr(RBox::new(1u8));

    assert_eq!(to.unchecked_downcast_as_mut::<u8>(), &mut 1u8);
}

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
P: GetPointerKind,
I: InterfaceType,

pub fn reborrow<'re>(&'re self) -> DynTrait<'borr, RRef<'re, ()>, I, EV> where
P: AsPtr<PtrTarget = ()>,
PrivStruct: ReborrowBounds<I::Send, I::Sync>,
EV: Copy,

Creates a shared reborrow of this DynTrait.

The reborrowed DynTrait cannot use these methods:

DynTrait::default

This is only callable if DynTrait is either Send + Sync or !Send + !Sync.

Example

use abi_stable::{
    erased_types::interfaces::DebugDisplayInterface,
    std_types::RBox,
    type_level::{impl_enum::Implemented, trait_marker},
    DynTrait, InterfaceType, RRef,
};

let to: DynTrait<'static, RBox<()>, DebugDisplayInterface> =
    DynTrait::from_value(1337_u16);

assert_eq!(debug_string(to.reborrow()), "1337");

fn debug_string<I>(to: DynTrait<'_, RRef<'_, ()>, I>) -> String
where
    I: InterfaceType<Debug = Implemented<trait_marker::Debug>>,
{
    format!("{:?}", to)
}

pub fn reborrow_mut<'re>(&'re mut self) -> DynTrait<'borr, RMut<'re, ()>, I, EV> where
P: AsMutPtr<PtrTarget = ()>,
PrivStruct: ReborrowBounds<I::Send, I::Sync>,
EV: Copy,

Creates a mutable reborrow of this DynTrait.

The reborrowed DynTrait cannot use these methods:

DynTrait::default

DynTrait::clone

This is only callable if DynTrait is either Send + Sync or !Send + !Sync.

Example

use abi_stable::{
    erased_types::interfaces::DEIteratorInterface, std_types::RBox, DynTrait,
};

let mut to = DynTrait::from_value(0_u8..=255).interface(DEIteratorInterface::NEW);

assert_eq!(both_ends(to.reborrow_mut()), (Some(0), Some(255)));
assert_eq!(both_ends(to.reborrow_mut()), (Some(1), Some(254)));
assert_eq!(both_ends(to.reborrow_mut()), (Some(2), Some(253)));
assert_eq!(both_ends(to.reborrow_mut()), (Some(3), Some(252)));

fn both_ends<I>(mut to: I) -> (Option<I::Item>, Option<I::Item>)
where
    I: DoubleEndedIterator,
{
    (to.next(), to.next_back())
}

impl<'borr, P, I, EV> DynTrait<'borr, P, I, EV>where
I: InterfaceType + 'borr,
EV: 'borr,
P: AsPtr,

pub fn default(&self) -> Selfwhere
P: AsPtr + GetPointerKind<Kind = PK_SmartPointer>,
I: InterfaceType<Default = Implemented<Default>>,
EV: Copy,

Constructs a DynTrait<P, I> with the default value for P.

Reborrowing

This cannot be called with a reborrowed DynTrait:

let object = DynTrait::from_value(()).interface(DefaultInterface);
let borrow = object.reborrow();
let _ = borrow.default();

let object = DynTrait::from_value(()).interface(DefaultInterface);
let borrow = object.reborrow_mut();
let _ = borrow.default();

Example

use abi_stable::{erased_types::interfaces::DebugDefEqInterface, DynTrait};

{
    let object = DynTrait::from_value(true).interface(DebugDefEqInterface);

    assert_eq!(
        object.default(),
        DynTrait::from_value(false).interface(DebugDefEqInterface)
    );
}
{
    let object = DynTrait::from_value(123u8).interface(DebugDefEqInterface);

    assert_eq!(
        object.default(),
        DynTrait::from_value(0u8).interface(DebugDefEqInterface)
    );
}

pub fn serialize_into_proxy<'a>(&'a self) -> Result<I::ProxyType, RBoxError>where
P: AsPtr,
I: InterfaceType<Serialize = Implemented<Serialize>> + GetSerializeProxyType<'a>,

It serializes a DynTrait<_> into a string by using <ConcreteType as SerializeType>::serialize_impl.

pub fn deserialize_from_proxy<'de>(proxy: I::Proxy) -> Result<Self, RBoxError>where
P: 'borr + AsPtr,
I: DeserializeDyn<'de, Self>,

Deserializes a DynTrait<'borr, _> from a proxy type, by using <I as DeserializeDyn<'borr, Self>>::deserialize_dyn.

impl<'borr, P, I, Item, EV> DynTrait<'borr, P, I, EV>where
P: AsMutPtr,
I: IteratorItemOrDefault<'borr, Item = Item> + InterfaceType<Iterator = Implemented<Iterator>>,
Item: 'borr,

pub fn skip_eager(&mut self, n: usize)

Eagerly skips n elements from the iterator.

This method is faster than using Iterator::skip.

Example

let mut iter = 0..20;
let mut wrapped = DynTrait::from_ptr(&mut iter).interface(IteratorInterface::NEW);

assert_eq!(wrapped.next(), Some(0));

wrapped.skip_eager(2);

assert_eq!(wrapped.next(), Some(3));
assert_eq!(wrapped.next(), Some(4));
assert_eq!(wrapped.next(), Some(5));

wrapped.skip_eager(2);

assert_eq!(wrapped.next(), Some(8));
assert_eq!(wrapped.next(), Some(9));

wrapped.skip_eager(9);

assert_eq!(wrapped.next(), Some(19));
assert_eq!(wrapped.next(), None);

pub fn extending_rvec(&mut self, buffer: &mut RVec<Item>, taking: ROption<usize>)

Extends the RVec<Item> with the self Iterator.

Extends buffer with as many elements of the iterator as taking specifies:

RNone: Yields all elements.Use this with care, since Iterators can be infinite.

RSome(n): Yields n elements.

Example

let mut wrapped = DynTrait::from_value(0..).interface(IteratorInterface::NEW);

let mut buffer = vec![101, 102, 103].into_c();
wrapped.extending_rvec(&mut buffer, RSome(5));
assert_eq!(&buffer[..], &*vec![101, 102, 103, 0, 1, 2, 3, 4]);

assert_eq!(wrapped.next(), Some(5));
assert_eq!(wrapped.next(), Some(6));
assert_eq!(wrapped.next(), Some(7));

impl<'borr, P, I, Item, EV> DynTrait<'borr, P, I, EV>where
Self: Iterator<Item = Item>,
P: AsMutPtr,
I: IteratorItemOrDefault<'borr, Item = Item> + InterfaceType<DoubleEndedIterator = Implemented<DoubleEndedIterator>>,
Item: 'borr,

pub fn nth_back_(&mut self, nth: usize) -> Option<Item>

Gets teh nth element from the back of the iterator.

Example

use abi_stable::{erased_types::interfaces::DEIteratorCloneInterface, DynTrait};

let to = || DynTrait::from_value(7..=10).interface(DEIteratorCloneInterface::NEW);

assert_eq!(to().nth_back_(0), Some(10));
assert_eq!(to().nth_back_(1), Some(9));
assert_eq!(to().nth_back_(2), Some(8));
assert_eq!(to().nth_back_(3), Some(7));
assert_eq!(to().nth_back_(4), None);
assert_eq!(to().nth_back_(5), None);

pub fn extending_rvec_back(
&mut self,
buffer: &mut RVec<Item>,
taking: ROption<usize>
)

Extends the RVec<Item> with the back of the self DoubleEndedIterator.

Extends buffer with as many elements of the iterator as taking specifies:

RNone: Yields all elements.Use this with care, since Iterators can be infinite.

RSome(n): Yields n elements.

Example

let mut wrapped = DynTrait::from_value(0..=3).interface(DEIteratorInterface::NEW);

let mut buffer = vec![101, 102, 103].into_c();
wrapped.extending_rvec_back(&mut buffer, RNone);
assert_eq!(&buffer[..], &*vec![101, 102, 103, 3, 2, 1, 0])