Module abi_stable::for_examples

Types used in documentation examples.

Structs

Module

This type is used in prefix type examples.

Module_Prefix

This is the prefix fields of Module, accessible through Module_Ref, with .0.prefix().

Module_Ref

This is the pointer to the prefix of Module.

PhantModule

This type is used in prefix type examples.

PhantModule_Prefix

This is the prefix fields of PhantModule< T >, accessible through PhantModule_Ref, with .0.prefix().

PhantModule_Ref

This is the pointer to the prefix of PhantModule< T >.

ValidTag_Interface

Describes the traits required when constructing a NonExhaustive<> from ValidTag,by implementing InterfaceType.

ValidTag_NEMarker

A marker type which implements StableAbi with the layout of ValidTag,used as a phantom field of NonExhaustive.

ValidTag_Storage

The default InlineStorage that NonExhaustive uses for ValidTag.

Enums

ValidTag

For demonstrating ffi-safe non-exhaustive enums.

Traits

ValidTag_Bounds

An alias for the traits that NonExhaustive<ValidTag,_,_> requires to be constructed,and implements afterwards.

Type Aliases

ValidTag_NE

An alias for NonExhaustive wrapping a ValidTag.