abi_stable

Module inline_storage

Contains the InlineStorage trait,and related items.

Modules

alignment

Helper types related to the alignemnt of inline storage.

Traits

AlignerFor

For getting the AlignTo* type which aligns Self to ALIGNMENT.

InlineStorage

Type used as the inline storage of a RSmallBox<>/NonExhaustive<>.

Type Aliases

GetAlignerFor

For getting the AlignTo* type which aligns T to ALIGNMENT.