abi_stable::library

Function abi_header_from_raw_library

pub unsafe fn abi_header_from_raw_library(
    raw_library: &RawLibrary,
) -> Result<AbiHeaderRef, LibraryError>

Gets the AbiHeaderRef of a library.

Errors

This will return these errors:

LibraryError::GetSymbolError: If the root module was not exported.

Safety

The AbiHeaderRef is implicitly tied to the lifetime of the library, it will contain dangling 'static references if the library is dropped before it does.