Function abi_stable::nonexhaustive_enum::assert_correct_storage

pub const fn assert_correct_storage<Enum, Storage>(args: AssertCsArgs)

Asserts that the size and alignment of an enum aree valid for this storage.

To make this function a const fn, the names of the Enum and Storage types must be passed separately.