Module abi_stable::sabi_trait::doc_examples::Doer_trait

This module is generated by the #[sabi_trait] attribute on Doer

Structs

Doer_Interface

A marker type describing the traits that are required when constructing Doer_TO,and are then implemented by it, by implementing the InterfaceType trait.

Doer_TO

The trait object for Doer.

Traits

Doer

An example trait object that uses RObject as a backend.

Type Aliases

Doer_Backend

An alias for the underlying implementation of Doer_TO.

Doer_CTO

A type alias for the const-constructible Doer_TO.