Module abi_stable::sabi_types::bitarray

A packed array of boolean enum values.

Structs

BitArray64

An array of 64 binary enum values.

BitArray64Iter

Iterator over the enums inside a BitArray64

Traits

BooleanEnum

A trait for enums with two variants where one is truthy and the other one is falsy.

Functions

bool_to_enum

Converts a bool to a BooleanEnum.

enum_to_bool

Converts a BooleanEnum to a bool