Trait abi_stable::sabi_types::bitarray::BooleanEnum

pub unsafe trait BooleanEnum: Debug + Copy + 'static {
    const FALSE: Self;
    const TRUE: Self;
}

A trait for enums with two variants where one is truthy and the other one is falsy.

Safety

This type must:

be represented as a u8

have exactly two possible values

assign one value to the FALSE associated constant,

assign another value to the TRUE associated constant

Required Associated Constants

const FALSE: Self

The falsy value of this type

const TRUE: Self

The truthy value of this type

Object Safety

This trait is not object safe.