abi_stable::std_types::cow

Enum RCow

#[repr(C)]
pub enum RCow<B, O> {
    Borrowed(B),
    Owned(O),
}

Ffi-safe equivalent of std::borrow::Cow.

This has type aliases for the three most common usecases:

RCowStr: contains an RStr<'_> or an RString.

RCowSlice: contains an RSlice<'_, T> or an RVec<T>.

RCowVal: contains a &T or a T.

Example

Using a RCowStr<'a>.

This implements a solution to the well known fizzbuzz problem.

use abi_stable::std_types::{RCow, RCowStr};

fn fizzbuzz(n: u32) -> RCowStr<'static> {
    match (n % 3, n % 5) {
        (0, 0) => RCow::from("FizzBuzz"),
        (0, _) => RCow::from("Fizz"),
        (_, 0) => RCow::from("Buzz"),
        (_, _) => RCow::from(n.to_string()),
    }
}

for n in 1..=100 {
    println!("{}", fizzbuzz(n));
}

Note: this example allocates when the number is neither a multiple of 5 or 3.

Variants

Borrowed(B)

Owned(O)

Implementations

impl<B> RCow<B, B::ROwned>
where B: IntoOwned,

pub fn to_mut(&mut self) -> &mut B::ROwned

Get a mutable reference to the owned form of RCow, converting to the owned form if it is currently the borrowed form.

Examples

use abi_stable::std_types::{RCow, RCowStr};

let mut cow: RCowStr<'_> = RCow::from("Hello");

assert_eq!(&*cow, "Hello");
assert!(cow.is_borrowed());

cow.to_mut().push_str(", world!");

assert!(cow.is_owned());
assert_eq!(cow, RCow::from("Hello, world!"));

pub fn into_owned(self) -> B::ROwned

Unwraps into the owned owner form of RCow, converting to the owned form if it is currently the borrowed form.

Examples

use abi_stable::std_types::{RCow, RCowStr};

let mut cow: RCowStr<'_> = RCow::from("Hello");

assert_eq!(&*cow, "Hello");

let mut buff = cow.into_owned();
buff.push_str(", world!");

assert_eq!(&*buff, "Hello, world!");

pub fn borrowed(&self) -> &<B as Deref>::Target

Gets the contents of the RCow casted to the borrowed variant.

Examples

use abi_stable::std_types::{RCow, RCowSlice, RSlice};
{
    let cow: RCowSlice<'_, u8> = RCow::from(&[0, 1, 2, 3][..]);
    assert_eq!(cow.borrowed(), RSlice::from_slice(&[0, 1, 2, 3]));
}
{
    let cow: RCowSlice<'_, u8> = RCow::from(vec![0, 1, 2, 3]);
    assert_eq!(cow.borrowed(), RSlice::from_slice(&[0, 1, 2, 3]));
}

impl<B, O> RCow<B, O>

pub const fn is_borrowed(&self) -> bool

Whether this is a borrowing RCow.

Examples

use abi_stable::std_types::{RCow, RCowSlice};

{
    let cow: RCowSlice<'_, u8> = RCow::from(&[0, 1, 2, 3][..]);
    assert!(cow.is_borrowed());
}
{
    let cow: RCowSlice<'_, u8> = RCow::from(vec![0, 1, 2, 3]);
    assert!(!cow.is_borrowed());
}

pub const fn is_owned(&self) -> bool

Whether this is an owning RCow.

Examples

use abi_stable::std_types::{RCow, RCowSlice};

let cow: RCowSlice<'_, u8> = RCow::from(&[0, 1, 2, 3][..]);
assert!(!cow.is_owned());

let cow: RCowSlice<'_, u8> = RCow::from(vec![0, 1, 2, 3]);
assert!(cow.is_owned());

impl<'a> RCow<RStr<'a>, RString>

pub const fn from_str(this: &'a str) -> Self

For converting a &'a [T] to an RCowSlice<'a, T>, most useful when converting from &'a [T;N] because it coerces the array to a slice.

Example

use abi_stable::std_types::{RCow, RCowStr};

const C: RCowStr<'_> = RCow::from_str("hello");

assert_eq!(C, "hello");

pub fn as_str(&self) -> &str

Borrows this RCow as a str.

Conditional const fn

This function requires the rust_1_64 feature to be const-callable

Example

use abi_stable::std_types::RCow;

let cow = RCow::from_str("world");

assert_eq!(cow.as_str(), "world")

impl<'a, T> RCow<RSlice<'a, T>, RVec<T>>

pub const fn from_slice(this: &'a [T]) -> Self

For converting a &'a [T] to an RCowSlice<'a, T>, most useful when converting from &'a [T;N] because it coerces the array to a slice.

Example

use abi_stable::std_types::{RCow, RCowSlice};

const C: RCowSlice<'_, u8> = RCow::from_slice(&[3, 5, 8]);

assert_eq!(C, [3, 5, 8]);

pub fn as_slice(&self) -> &[T]

Borrows this RCow as a slice.

Conditional const fn

This function requires the rust_1_64 feature to be const-callable

Example

use abi_stable::std_types::RCow;

let cow = RCow::from_slice(&[3, 5, 8]);

assert_eq!(cow.as_slice(), [3, 5, 8])