abi_stable::std_types::vec

Struct Drain

#[repr(C)]
pub struct Drain<'a, T> { /* private fields */ }

An Iterator returned by RVec::drain , which removes and yields all the elements in a range from the RVec<T>.

Implementations

impl<'a, T> Drain<'a, T>

pub fn as_slice(&self) -> &[T]

Returns a slice over the remainder of the Vec<T> that is being drained.

Example

use abi_stable::std_types::RVec;

let mut list = (0..8).collect::<RVec<u8>>();
let mut iter = list.drain(3..7);

assert_eq!(iter.as_slice(), &[3, 4, 5, 6]);

assert_eq!(iter.next(), Some(3));
assert_eq!(iter.as_slice(), &[4, 5, 6]);

assert_eq!(iter.next(), Some(4));
assert_eq!(iter.as_slice(), &[5, 6]);

drop(iter);

assert_eq!(list.as_slice(), &[0, 1, 2, 7]);

pub fn as_mut_slice(&mut self) -> &mut [T]

Returns a mutable slice over the remainder of the Vec<T> that is being drained.

Example

use abi_stable::std_types::RVec;

let mut list = (0..8).collect::<RVec<u8>>();
let mut iter = list.drain(3..7);

assert_eq!(iter.as_mut_slice(), &mut [3, 4, 5, 6]);

assert_eq!(iter.next(), Some(3));
assert_eq!(iter.as_mut_slice(), &mut [4, 5, 6]);

assert_eq!(iter.next(), Some(4));
assert_eq!(iter.as_mut_slice(), &mut [5, 6]);

drop(iter);

assert_eq!(list.as_slice(), &[0, 1, 2, 7]);