Function abi_stable::utils::leak_value

pub fn leak_value<'a, T>(value: T) -> &'a T
where
    T: 'a,

Leaks value into the heap, and returns a reference to it.

Warning

You must be careful when calling this function, since this leak is ignored by miri.