Module abi_stable::external_types

Ffi wrapper for types defined outside the standard library.

The modules here are named after the crates whose types are being wrapped.

Re-exports

pub use self::parking_lot::RMutex;

pub use self::parking_lot::ROnce;

pub use self::parking_lot::RRwLock;

pub use self::serde_json::RawValueBox;

pub use self::serde_json::RawValueRef;

Modules

crossbeam_channel

Ffi-safe wrapper types around the crossbeam-channel channel types.

parking_lot

Ffi-safe synchronization primitives,most of which are ffi-safe wrappers of parking_lot types

serde_json

Ffi-safe equivalents of serde_json types.