abi_stable::external_types

Module crossbeam_channel

Ffi-safe wrapper types around the crossbeam-channel channel types.

Structs

RIntoIter

An iterator which receives the values sent through the channel, blocking until a value is received.

RIter

An iterator which receives the values sent through the channel, blocking until a value is received.

RReceiver

The receiver end of a channel, which can be either bounded or unbounded.

RSender

The sender end of a channel, which can be either bounded or unbounded.

Functions

bounded

Creates a channel which can hold up to capacity elements in its internal queue.

never

Creates a receiver that can never receive any value.

unbounded

Creates a channel which can hold an unbounded amount elements in its internal queue.