abi_stable::inline_storage

Module alignment

Helper types related to the alignemnt of inline storage.

Structs

AlignTo1

Aligns its contents to an address at a multiple of 1 bytes.

AlignTo2

Aligns its contents to an address at a multiple of 2 bytes.

AlignTo4

Aligns its contents to an address at a multiple of 4 bytes.

AlignTo8

Aligns its contents to an address at a multiple of 8 bytes.

AlignTo16

Aligns its contents to an address at a multiple of 16 bytes.

AlignTo32

Aligns its contents to an address at a multiple of 32 bytes.

AlignTo64

Aligns its contents to an address at a multiple of 64 bytes.

AlignTo128

Aligns its contents to an address at a multiple of 128 bytes.

AlignTo256

Aligns its contents to an address at a multiple of 256 bytes.

AlignTo512

Aligns its contents to an address at a multiple of 512 bytes.

AlignTo1024

Aligns its contents to an address at a multiple of 1024 bytes.

AlignTo2048

Aligns its contents to an address at a multiple of 2048 bytes.

AlignTo4096

Aligns its contents to an address at a multiple of 4096 bytes.

AlignTo8192

Aligns its contents to an address at a multiple of 8192 bytes.

AlignTo16384

Aligns its contents to an address at a multiple of 16384 bytes.

AlignTo32768

Aligns its contents to an address at a multiple of 32768 bytes.

AlignToUsize

Aligns its contents to an address to an address at a multiple of the size of a pointer.