abi_stable

Module marker_type

Zero-sized types .

Structs

ErasedObject

Used by vtables/pointers to signal that the type has been erased.

ErasedPrefix

Used by pointers to vtables/modules to signal that the type has been erased.

NonOwningPhantom

An ffi-safe equivalent of a PhantomData<fn()->T>

NotCopyNotClone

Zero-sized marker type used to signal that even though a type could implement Copy and Clone, it is semantically an error to do so.

SyncSend

Marker type used to mark a type as being Send + Sync.

SyncUnsend

Marker type used to mark a type as being !Send + Sync.

UnsafeIgnoredType

MarkerType which ignores its type parameter in its StableAbi implementation.

UnsyncSend

Marker type used to mark a type as being Send + !Sync.

UnsyncUnsend

Marker type used to mark a type as being !Send + !Sync.