Trait abi_stable::nonexhaustive_enum::NonExhaustiveMarker

pub unsafe trait NonExhaustiveMarker<Storage>: GetEnumInfo {
    type Marker;
}

Queries the marker type which describes the layout of this enum, for use in NonExhaustive’s StableAbi impl.

Safety

Self::Marker must describe the layout of this enum, with the size and alignment of Storage, and using IsExhaustive::nonexhaustive to construct IsExhaustive in the enum’s TypeLayout.

Required Associated Types

type Marker

A marker type which describes the layout of this enum in its StableAbi impl.