abi_stable::nonexhaustive_enum

Trait NonExhaustiveMarker

pub unsafe trait NonExhaustiveMarker<Storage>: GetEnumInfo {
    type Marker;
}

Queries the marker type which describes the layout of this enum, for use in NonExhaustive’s StableAbi impl.

Safety

Self::Marker must describe the layout of this enum, with the size and alignment of Storage, and using IsExhaustive::nonexhaustive to construct IsExhaustive in the enum’s TypeLayout.

Required Associated Types

type Marker

A marker type which describes the layout of this enum in its StableAbi impl.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.