abi_stable

Module prefix_type

Types,traits,and functions used by prefix-types.

Structs

PTStructLayout

Represents the layout of a prefix-type,for use in error messages.

PrefixRef

A reference to a prefix type.

WithMetadata_

Wraps a type along with its prefix-type-related metadata, so that it can be converted to its prefix.

Enums

IsAccessible

Whether a field is accessible.

IsConditional

Whether a field is conditional, whether it has a #[sabi(accessible_if = expression)] helper attribute or not.

Traits

PrefixRefTrait

Marker trait for pointers to prefix field structs.

PrefixTypeTrait

For types deriving StableAbi with #[sabi(kind(Prefix(..)))].

Functions

panic_on_missing_field_ty

Used to panic with an error message informing the user that a field is expected to be on the T type when it’s not.

panic_on_missing_fieldname

Used to panic with an error message informing the user that a field is expected to be on the T type when it’s not.

Type Aliases

FieldAccessibility

An array with whether the ith field of a prefix-type is accessible through its accessor method.

FieldConditionality

An array with whether the ith field in the prefix of a prefix-type is conditional,which means whether it has the #[sabi(accessible_if = expression)] attribute applied to it.

WithMetadata

Alias for WithMetadata_ that defaults to passing <T as PrefixTypeTrait>::PrefixFields as the second type parameter.