abi_stable::sabi_trait::doc_examples

Module Action_trait

This module is generated by the #[sabi_trait] attribute on Action

Structs

Action_Interface

A marker type describing the traits that are required when constructing Action_TO,and are then implemented by it, by implementing the InterfaceType trait.

Action_TO

The trait object for Action.

Traits

Action

The trait used in examples of #[sabi_trait] trait object methods, in abi_stable::docs::sabi_trait_inherent

Type Aliases

Action_Backend

An alias for the underlying implementation of Action_TO.

Action_CTO

A type alias for the const-constructible Action_TO.