Module abi_stable::sabi_types

ffi-safe types that aren’t wrappers for other types.

Re-exports

pub use self::bitarray::BitArray64;

pub use self::rsmallbox::RSmallBox;

pub use self::version::ParseVersionError;

pub use self::version::VersionNumber;

pub use self::version::VersionStrings;

Modules

bitarray

A packed array of boolean enum values.

rsmallbox

Contains the RSmallBox<_> type.

version

Types representing the version number of a library.

Structs

CmpIgnored

Wrapper type used to ignore its contents in comparisons.

Constructor

Newtype wrapper to pass function pointers to const fn.

LateStaticRef

A late-initialized static reference,with fallible initialization.

MovePtr

A move pointer, which allows moving the value from the reference, consuming it in the process.

NulStr

A utf8 nul-terminated immutable borrowed string.

RMut

Equivalent to &'a mut T, which allows a few more operations without causing Undefined Behavior.

RRef

Equivalent to &'a T, which allows a few more operations without causing Undefined Behavior.

StaticRef

A wrapper type for vtable static references, and other constants that have non-'static generic parameters but are safe to reference for the lifetime of T.

Enums

ConstructorOrValue

Either the constructor for a value or the value itself

MaybeCmp

An Option-like type which only compares equal if it contains a value(the Just variant).

NulStrError

Error from trying to convert a &str to a NulStr