Module abi_stable::std_types::cow

Contains the ffi-safe equivalent of std::borrow::Cow, and related items.

Structs

BorrowingRCowStr

A helper type, to deserialize a RCowStr<'a> which borrows from the deserializer.

BorrowingRCowU8Slice

A helper type, to deserialize an RCow<'a, [u8]> which borrows from the deserializer.

Enums

RCow

Ffi-safe equivalent of std::borrow::Cow.

Traits

RCowCompatibleRef

For making a Cow<'a, Self> convertible into an RCow.

Functions

deserialize_borrowed_bytes

Deserializes an RCow<'a, [u8]> that borrows the slice from the deserializer whenever possible.

deserialize_borrowed_str

Deserializes an RCowStr<'a> that borrows the string from the deserializer whenever possible.

Type Aliases

RCowSlice

Ffi-safe equivalent of Cow<'a, [T]>, either an RSlice or RVec.

RCowStr

Ffi-safe equivalent of Cow<'a, str>, either an RStr or RString.

RCowVal

Ffi-safe equivalent of Cow<'a, T>, either a &T or T.