Module abi_stable::std_types::map

Contains the ffi-safe equivalent of std::collections::HashMap, and related items.

Structs

IntoIter

An iterator that yields all the entries of an RHashMap, deallocating the hashmap afterwards.

Keys

An iterator over the keys of a RHashMap.

MutIterInterface

The InterfaceType of the IterMut iterator for RHashMap.

RHashMap

An ffi-safe hashmap, which wraps std::collections::HashMap<K, V, S>, only requiring the K: Eq + Hash bounds when constructing it.

ROccupiedEntry

A handle into an occupied entry in a map.

RVacantEntry

A handle into a vacant entry in a map.

RefIterInterface

The InterfaceType of the Iter iterator for RHashMap.

ValIterInterface

The InterfaceType of the Drain iterator for RHashMap.

Values

An iterator over the values of a HashMap.

Enums

REntry

A handle into an entry in a map, which is either vacant or occupied.

Type Definitions

Drain

An RHashMap iterator, implementing Iterator<Item= Tuple2< K, V > > + !Send + !Sync

Iter

An RHashMap iterator, implementing Iterator<Item= Tuple2< &K, &V > > + !Send + !Sync + Clone

IterMut

An RHashMap iterator, implementing Iterator<Item= Tuple2< &K, &mut V > > + !Send + !Sync