abi_stable::type_level
Module downcasting

Type-level enum representing whether a DynTrait/RObject/#[sabi_trait]-generated trait object can be converted back into the concrete type they were constructed with.

Structs

TD_CanDowncast

Passed to trait object constructors to make the trait object downcast capable, as opposed to TD_Opaque.

TD_Opaque

Passed to trait object constructors to make it impossible to downcast the trait object, as opposed to TD_CanDowncast.

Traits

GetUTID

Gets a function optionally returning the UTypeId of T.