Module abi_stable::type_level::impl_enum

Type-level-enum representing whether a trait is implemented or not implemented.

Structs

Implemented

Describes that a trait must be implemented.

Unimplemented

Describes that a trait does not need to be implemented.

Traits

ImplFrom_

Converts a type to either Unimplemented or Implemented.

Implementability

Trait for Implemented and Unimplemented

Type Aliases

ImplFrom

Converts B to either Unimplemented<T> or Implemented<T>.